requirements.txt
crewai

python-dotenv
nanda-adapter
requests